// src/api/generated/model/getAllImportJobsStatus.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */
export type GetAllImportJobsStatus = typeof GetAllImportJobsStatus[keyof typeof GetAllImportJobsStatus];


export const GetAllImportJobsStatus = {
  PENDING: 'PENDING',
  RUNNING: 'RUNNING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
} as const;