// src/api/generated/model/importJobSummaryDto.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */import type { ImportJobSummaryDtoStatus } from './importJobSummaryDtoStatus';
import type { ImportJobSummaryDtoType } from './importJobSummaryDtoType';

export interface ImportJobSummaryDto {
  importJobId: string;
  type: ImportJobSummaryDtoType;
  status: ImportJobSummaryDtoStatus;
  fileName: string;
  readCount: number;
  writeCount: number;
  skipCount: number;
  startedAt: string;
  finishedAt: string;
  errorMessage: string;
  createdAt: string;
}