// src/api/generated/model/import-job-summary-dto.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */import type { ImportJobSummaryDtoStatus } from './import-job-summary-dto-status';
import type { ImportJobSummaryDtoType } from './import-job-summary-dto-type';

export interface ImportJobSummaryDto {
  importJobId: string;
  type: ImportJobSummaryDtoType;
  status: ImportJobSummaryDtoStatus;
  fileName: string;
  readCount: number;
  writeCount: number;
  skipCount: number;
  startedAt: string;
  finishedAt: string;
  errorMessage: string;
  createdAt: string;
}