// src/api/generated/model/pageDtoImportJobSummaryDto.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */import type { ImportJobSummaryDto } from './importJobSummaryDto';
import type { PaginationMeta } from './paginationMeta';

export interface PageDtoImportJobSummaryDto {
  content: ImportJobSummaryDto[];
  meta: PaginationMeta;
}