// src/api/generated/model/page-dto-import-job-summary-dto.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */import type { ImportJobSummaryDto } from './import-job-summary-dto';
import type { PaginationMeta } from './pagination-meta';

export interface PageDtoImportJobSummaryDto {
  content: ImportJobSummaryDto[];
  meta: PaginationMeta;
}